requirements.txt
astroid==1.6.6
autopep8==1.5.4
backports.functools-lru-cache==1.6.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==4.0.2
enum34==1.1.10
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
futures==3.3.0
gunicorn==19.10.0
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.6
pandas==0.24.2
pycodestyle==2.6.0
pylint==1.9.5
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
